add feature to load external modules

+++ Zect/modules/paste.py
@@ -4,9 +4,10 @@
 # and is released under the "AGP v3.0 License Agreement".
 # Please see < https://github.com/shre-yansh/ZectUserBot/blob/master/LICENSE >
 #
-# All rights reserved.
 
+        
 import asyncio
+import os
 import requests
 import aiohttp
 from pyrogram import filters
@@ -25,43 +26,44 @@ CMD_HELP.update(
     }
 )
 
-
+def paste(text):
+    try:
+        data = {"content": text, "extension": "txt"}
+        url = "https://spaceb.in/api/v1/documents"
+        neko = requests.post(url, data=data)
+        data = neko.json()
+    except Exception as e:
+        return False
+    else:
+        id = data["payload"]["id"]
+        # id = f'https://spaceb.in/{data["payload"]["id"]}'
+        return id
 @app.on_message(
-    filters.command(["neko", "paste"], PREFIX) & filters.me
+    filters.command(["paste"], PREFIX) & filters.me
 )
 async def neko(_, message: Message):
-    text = message.reply_to_message.text
-    try:
-        params = {"content": text}
-        headers = {'content-type' : 'application/json'}
-        url = "https://nekobin.com/api/documents"
-        neko = requests.post(url, json=params, headers=headers)
-        key = neko.json()["result"]["key"]
-    except Exception:
+    if message.reply_to_message.text:
+        text = message.reply_to_message.text
+    elif message.reply_to_message.document:
+        file = await app.download_media(message.reply_to_message.document.file_id)
+        with open(file, "r") as f:
+            text = f.read()
+        os.remove(file)
+    else:
+        await message.edit_text("`Reply to a text or document file`")
+        return
+    id = paste(text)
+    url = f'https://spaceb.in/{id}'
+    raw_url = f'https://spaceb.in/api/v1/documents/{id}/raw'
+    if not url:
         await message.edit_text("`API is down try again later`")
         await asyncio.sleep(2)
         await message.delete()
         return
     else:
-        url = f"https://nekobin.com/{key}"
-        reply_text = f"**Pasted to: [Nekobin]({url})**"
-        delete = (
-            True
-            if len(message.command) > 1
-            and message.command[1] in ["d", "del"]
-            and message.reply_to_message.from_user.is_self
-            else False
+        reply_text = f"**Pasted to: [Sbacebin]({url})\nRaw Url: [Raw]({raw_url})**"
+       
+        await message.edit_text(
+            reply_text,
+            disable_web_page_preview=True,
         )
-        if delete:
-            await asyncio.gather(
-                app.send_message(
-                    message.chat.id, reply_text, disable_web_page_preview=True
-                ),
-                message.reply_to_message.delete(),
-                message.delete(),
-            )
-        else:
-            await message.edit_text(
-                reply_text,
-                disable_web_page_preview=True,
-            )
